Tasks_and_codes/Day_16_20/requirements.txt
agno
groq
python-dotenv
pandas
numpy
ddgs
duckduckgo-search
yfinance
pypdf
lancedb
streamlit
tantivy 
openai 
sqlalchemy
chromadb